+++ CHANGELOG.md
@@ -131,7 +131,13 @@
     - air-primary, air-primary-success, air-primary-alert, air-primary-copilot, air-primary-warning
   - @deprecate color
     - default, danger, success, warning, primary, secondary, collab, ai
-* Checkbox
+* **Checkbox**
+  - fix color 
+  - add color
+    - air-primary, air-primary-success, air-primary-alert, air-primary-copilot, air-primary-warning
+  - @deprecate color
+    - default, danger, success, warning, primary, secondary, collab, ai
+* RadioGroup
   - fix color 
   - add color
     - air-primary, air-primary-success, air-primary-alert, air-primary-copilot, air-primary-warning